pyproject.toml
[project]
name = "game-theory-feature-selection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"

dependencies = [
    "Boruta",
    "feature-engine",
    "lightgbm",
    "pre-commit",
    "pyHSICLasso",
    "ruff",
    "scikit-fingerprints",
    "sage-importance",
    "shap",
]
